Add colors to ConsoleLogger :)
